[GIE/GLogue] rename Catalogue as GLogue
