+++ interactive_engine/executor/ir/core/src/lib.rs
@@ -17,8 +17,8 @@ use std::io;
 
 pub use crate::plan::ffi::*;
 
-pub mod catalogue;
 pub mod error;
+pub mod glogue;
 pub mod plan;
 
 #[macro_use]
