+++ interactive_engine/executor/ir/core/tests/common/pattern_cases.rs
@@ -18,9 +18,9 @@ use std::collections::{BTreeSet, HashMap};
 use ir_common::generated::algebra as pb;
 use ir_common::generated::common as common_pb;
 use ir_common::KeyId;
-use ir_core::catalogue::error::IrPatternResult;
-use ir_core::catalogue::pattern::*;
-use ir_core::catalogue::{PatternId, PatternLabelId};
+use ir_core::glogue::error::IrPatternResult;
+use ir_core::glogue::pattern::*;
+use ir_core::glogue::{PatternId, PatternLabelId};
 use ir_core::plan::meta::{PlanMeta, TagId};
 use rand::rngs::StdRng;
 use rand::seq::SliceRandom;
